Make tmp file to mark whether node creation has been done  (#190)
If a service terminates between running subprocess and sending the node creation message, items can get missed from star files.

This creates a .tmp file which marks whether node creation has happened, and only skips it it if the output exists and no tmp file is present.

Also removes cryolo cbox file on reruns as this causes problems.

+++ src/cryoemservices/services/topaz_pick.py
@@ -89,11 +89,16 @@ class TopazPick(CommonService):
             return
 
         # Check if this file has been run before
-        if Path(topaz_params.output_path).is_file():
+        if (
+            Path(topaz_params.output_path).is_file()
+            and not Path(topaz_params.output_path).with_suffix(".tmp").is_file()
+        ):
             job_is_rerun = True
             Path(topaz_params.output_path).unlink()
         else:
             job_is_rerun = False
+            Path(topaz_params.output_path).parent.mkdir(parents=True, exist_ok=True)
+            Path(topaz_params.output_path).with_suffix(".tmp").touch(exist_ok=True)
 
         # Determine the project directory or job directory
         job_num_search = re.search("/job[0-9]+/", topaz_params.output_path)
@@ -103,7 +108,6 @@ class TopazPick(CommonService):
             self.log.warning(f"Invalid job directory in {topaz_params.output_path}")
             rw.transport.nack(header)
             return
-        Path(topaz_params.output_path).parent.mkdir(parents=True, exist_ok=True)
 
         # Construct a command to run topaz with the given parameters
         self.log.info(
@@ -190,6 +194,8 @@ class TopazPick(CommonService):
         if not job_is_rerun:
             # Only do the node creator inserts for new files
             rw.send_to("node_creator", node_creator_parameters)
+            # Remove tmp file after requesting node creation
+            Path(topaz_params.output_path).with_suffix(".tmp").unlink(missing_ok=True)
 
         # Read picks for images service
         try:
